fix: check App object type instead of relying on the event hint
This allows us to achieve compatibility with all XF releases (the event hint broke in 2.3 because controller class names now have a `Controller` suffix) and also hardens the check. Previously you could have had the ability to trigger an error by navigating to `https://example.com///admin.php` and it would cause an error that would not be caught by the event listener, since the login page was rendered by the 'XF:Error' controller.